+++ src/isa/csrrc.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/csrrc.h"
 
-CSRRC::CSRRC(Register rd, Register rs1, uint32_t imm)
+CSRRC::CSRRC(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::CSRRC, rd, rs1, imm) {}
