+++ src/isa/csrrci.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/csrrci.h"
 
-CSRRCI::CSRRCI(Register rd, Register rs1, uint32_t imm)
+CSRRCI::CSRRCI(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::CSRRCI, rd, rs1, imm) {}
