+++ src/isa/csrrs.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/csrrs.h"
 
-CSRRS::CSRRS(Register rd, Register rs1, uint32_t imm)
+CSRRS::CSRRS(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::CSRRS, rd, rs1, imm) {}
