+++ src/isa/csrrwi.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/csrrwi.h"
 
-CSRRWI::CSRRWI(Register rd, Register rs1, uint32_t imm)
+CSRRWI::CSRRWI(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::CSRRWI, rd, rs1, imm) {}
