+++ src/isa/slli.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/slli.h"
 
-SLLI::SLLI(Register rd, Register rs1, uint32_t imm)
+SLLI::SLLI(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::SLLI, rd, rs1, imm) {}
