+++ src/isa/slliw.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/slliw.h"
 
-SLLIW::SLLIW(Register rd, Register rs1, uint32_t imm)
+SLLIW::SLLIW(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::SLLIW, rd, rs1, imm) {}
