+++ src/isa/slti.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/slti.h"
 
-SLTI::SLTI(Register rd, Register rs1, uint32_t imm)
+SLTI::SLTI(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::SLTI, rd, rs1, imm) {}
