+++ src/isa/sltiu.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/sltiu.h"
 
-SLTIU::SLTIU(Register rd, Register rs1, uint32_t imm)
+SLTIU::SLTIU(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::SLTIU, rd, rs1, imm) {}
