+++ src/isa/srliw.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/srliw.h"
 
-SRLIW::SRLIW(Register rd, Register rs1, uint32_t imm)
+SRLIW::SRLIW(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::SRLIW, rd, rs1, imm) {}
