+++ src/isa/xori.cc
@@ -1,4 +1,4 @@
 #include "rviss/isa/xori.h"
 
-XORI::XORI(Register rd, Register rs1, uint32_t imm)
+XORI::XORI(Register rd, Register rs1, uint16_t imm)
     : InstructionTypeI(Opcode::XORI, rd, rs1, imm) {}
